feat: CSRF protection added

+++ Http/models/Note.php
@@ -3,6 +3,7 @@
 namespace Http\Models;
 
 use App\Middlewares\Auth;
+use Core\Session;
 
 class Note
 {
@@ -10,6 +11,25 @@ class Note
     public static function create($connection, $noteData = [])
     {
         header('Content-Type: application/json');
+
+        if (!Session::validateCsrf($noteData['csrf_token'])) {
+            http_response_code(403);
+            echo json_encode([
+                'success' => false,
+                'data' => 'Invalid CSRF Token'
+            ]);
+            exit();
+        }
+
+        if (!Auth::user()) {
+            http_response_code(403);
+            echo json_encode([
+                'success' => false,
+                'data' => 'Unauthorized. You must login.'
+            ]);
+            exit();
+        }
+
         $sql = "INSERT INTO notes (name, priority, date, user_id, body) VALUES (:name, :priority, :date, :user_id, :body)";
         $note = $connection->insert($sql, [
             'name' => htmlspecialchars($noteData['name']),
@@ -18,19 +38,43 @@ class Note
             'user_id' => Auth::user(),
             'body' => htmlspecialchars($noteData['body']),
         ]);
+        http_response_code(201);
         echo json_encode([
             'success' => $note ? true : false,
             'data' => $note,
         ]);
+
+        Session::renewCsrf();
+
         exit();
     }
 
-    public static function get($connection, $id = null)
+    public static function get($connection, $id = null, $token)
     {
+
+        if (!Session::validateCsrf($token)) {
+            http_response_code(403);
+            echo json_encode([
+                'success' => false,
+                'data' => 'Invalid CSRF Token'
+            ]);
+            exit();
+        }
+
+        if (!Auth::user()) {
+            http_response_code(403);
+            echo json_encode([
+                'success' => false,
+                'data' => 'Unauthorized. You must login.'
+            ]);
+            exit();
+        }
 
         $sql = "SELECT * FROM notes WHERE user_id = :user_id";
 
         if (!Auth::isAdmin()) {
+            $sql .= " AND deleted_at NULL";
+        } else {
             $sql .= " AND deleted_at NOT NULL";
         }
 
@@ -54,6 +98,10 @@ class Note
             'user_id' => Auth::user(),
         ])->getAll();
 
+        http_response_code(200);
+
+        Session::renewCsrf();
+
         return json_encode([
             'success' => $note ? true : false,
             'data' => $note,
@@ -64,6 +112,24 @@ class Note
     {
         header('Content-Type: application/json');
 
+        if (!Session::validateCsrf($noteData['csrf_token'])) {
+            http_response_code(403);
+            echo json_encode([
+                'success' => false,
+                'data' => 'Invalid CSRF Token'
+            ]);
+            exit();
+        }
+
+        if (!Auth::user()) {
+            http_response_code(403);
+            echo json_encode([
+                'success' => false,
+                'data' => 'Unauthorized. You must login.'
+            ]);
+            exit();
+        }
+
         $sql = "UPDATE notes SET name = :name, priority = :priority, date = :date, body = :body WHERE id = :id AND user_id = :user_id";
         $note = $connection->update($sql, [
             'name' => htmlspecialchars($noteData['name']),
@@ -73,28 +139,53 @@ class Note
             'id' => $noteData['id'],
             'user_id' => Auth::user(),
         ]);
-
+        http_response_code(201);
         echo json_encode([
             'success' => $note ? true : false,
             'data' => $note,
         ]);
+
+        Session::renewCsrf();
+
         exit();
     }
 
-    public static function delete($connection, $id)
+    public static function delete($connection, $id, $token)
     {
         header('Content-Type: application/json');
+
+        if (!Session::validateCsrf($token)) {
+            http_response_code(403);
+            echo json_encode([
+                'success' => false,
+                'data' => 'Invalid CSRF Token'
+            ]);
+            exit();
+        }
+
+        if (!Auth::user()) {
+            http_response_code(403);
+            echo json_encode([
+                'success' => false,
+                'data' => 'Unauthorized. You must login.'
+            ]);
+            exit();
+        }
+
         $sql = "UPDATE notes SET deleted_at = :date WHERE id = :id AND user_id = :user_id";
         $note = $connection->delete($sql, [
             'date' => date_create(),
             'id' => $id,
             'user_id' => Auth::user(),
         ]);
-
+        http_response_code(200);
         echo json_encode([
             'success' => $note ? true : false,
             'data' => $note,
         ]);
+
+        Session::renewCsrf();
+
         exit();
     }
 }
